requirements.txt
transformers
datasets
torch
numpy
pandas
matplotlib
hydra-core
mlflow
wandb
tqdm 
h5py 
PyYAML 
tokenizers 